Moved qc.model as model at top-level. Now we have bnm.recon.model directory.
