+++ bnm/recon/algo/service/volume.py
@@ -5,7 +5,7 @@ import numpy
 import scipy.ndimage
 from bnm.recon.algo.service.annotation import AnnotationService
 from bnm.recon.io.volume import VolumeIO
-from bnm.recon.qc.model.volume import Volume
+from bnm.recon.model.volume import Volume
 from scipy.spatial.distance import pdist, squareform
