+++ bnm/recon/qc/image/processor.py
@@ -7,9 +7,9 @@ from bnm.recon.io.generic import GenericIO
 from bnm.recon.io.surface import FreesurferIO, GiftiSurfaceIO
 from bnm.recon.io.volume import VolumeIO
 from bnm.recon.logger import get_logger
-from bnm.recon.qc.image.writer import ImageWriter
-from bnm.recon.qc.model.constants import PROJECTIONS, SNAPSHOT_NAME, GIFTI_EXTENSION, T1_RAS_VOLUME, MRI_DIRECTORY, \
+from bnm.recon.model.constants import PROJECTIONS, SNAPSHOT_NAME, GIFTI_EXTENSION, T1_RAS_VOLUME, MRI_DIRECTORY, \
     FS_TO_CONN_INDICES_MAPPING_PATH
+from bnm.recon.qc.image.writer import ImageWriter
 
 
 class ImageProcessor(object):
