+++ bnm/recon/qc/surface_transform.py
@@ -5,8 +5,8 @@ import os
 import numpy
 from bnm.recon.io.generic import GenericIO
 from bnm.recon.logger import get_logger
+from bnm.recon.model.constants import SNAPSHOTS_DIRECTORY_ENVIRON_VAR, SNAPSHOT_NUMBER_ENVIRON_VAR
 from bnm.recon.qc.image.processor import ImageProcessor
-from bnm.recon.qc.model.constants import SNAPSHOTS_DIRECTORY_ENVIRON_VAR, SNAPSHOT_NUMBER_ENVIRON_VAR
 
 
 def parse_arguments():
